+++ src/backend/middleware/asyncRoute.js
@@ -1,0 +1,9 @@
+'use strict';
+
+function asyncRoute(handler) {
+  return (req, res, next) => {
+    Promise.resolve(handler(req, res, next)).catch(next);
+  };
+}
+
+module.exports = asyncRoute;
